new map applied to 2/3 vies in external session.
